--- src/client/src/components/AlternativeBlockForm/blockSchema.ts
@@ -1,17 +1,0 @@
-import { TFunction } from "i18next";
-import { DESCRIPTION_LENGTH, NAME_LENGTH, NOTATION_LENGTH } from "types/common/stringLengthConstants";
-import * as yup from "yup";
-
-export const blockSchema = (t: TFunction<"translation">) =>
-  yup.object({
-    name: yup
-      .string()
-      .max(NAME_LENGTH, t("common.validation.name.max", { length: NAME_LENGTH }))
-      .required(t("common.validation.name.required")),
-
-    description: yup
-      .string()
-      .max(DESCRIPTION_LENGTH, t("common.validation.description.max", { length: DESCRIPTION_LENGTH })),
-
-    notation: yup.string().max(NOTATION_LENGTH, t("common.validation.notation.max", { length: NOTATION_LENGTH })),
-  });
