--- src/client/src/components/AlternativeBlockForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AlternativeBlockForm";
