+++ src/client/src/components/BlockForm/AttributesStep.tsx
