+++ src/client/src/components/BlockForm/SelectSymbolStep.tsx
