+++ src/client/src/components/BlockForm/SymbolCard.tsx
