+++ src/client/src/components/BlockForm/TerminalsStep.helpers.ts
