+++ src/client/src/components/BlockForm/TerminalsStep.tsx
