+++ src/client/src/helpers/form.helpers.ts
@@ -2,6 +2,7 @@ import { toast } from "@mimirorg/component-library";
 import { UseQueryResult } from "@tanstack/react-query";
 import { AxiosError } from "axios";
 import { useEffect, useState } from "react";
+import { DefaultValues, KeepStateOptions } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 
 export const onSubmitForm = <TAm, TCm>(
@@ -22,6 +23,24 @@ export const onSubmitForm = <TAm, TCm>(
  * @param populateForm populates the form with the provided data
  */
 export const usePrefilledForm = <TIn, TOut>(
+  query: UseQueryResult<TIn>,
+  mapQueryDataToFormModel: (data: TIn) => TOut,
+  populateForm: (values?: DefaultValues<TOut> | TOut, keepStateOptions?: KeepStateOptions) => void,
+): [isPrefilled: boolean, isLoading: boolean] => {
+  const [isPrefilled, setIsPrefilled] = useState(false);
+
+  useEffect(() => {
+    if (!isPrefilled && query.isSuccess) {
+      setIsPrefilled(true);
+      populateForm(mapQueryDataToFormModel(query.data), { keepDefaultValues: false });
+    }
+  }, [query.data, query.isSuccess, populateForm, isPrefilled, mapQueryDataToFormModel]);
+
+  return [isPrefilled, query.isInitialLoading];
+};
+
+// This function is temporarily added while forms are in different states (using react hook form or not)
+export const usePrefilledFormTemporary = <TIn, TOut>(
   query: UseQueryResult<TIn>,
   mapQueryDataToFormModel: (data: TIn) => TOut,
   populateForm: (values: TOut) => void,
