+++ src/client/src/types/common/hasCardinality.ts
@@ -1,4 +1,4 @@
 export interface HasCardinality {
   minCount: number;
-  maxCount: number | null;
+  maxCount?: number | null;
 }
